Refactor: Create Plan class
