+++ frontend/.editorconfig
@@ -862,12 +862,13 @@ ij_html_space_inside_empty_tag = false
 ij_html_text_wrap = normal
 
 [{*.kt,*.kts}]
-ij_kotlin_align_in_columns_case_branch = false
+ij_any_call_parameters_wrap = on_every_item
+ij_kotlin_align_in_columns_case_branch = true
 ij_kotlin_align_multiline_binary_operation = false
-ij_kotlin_align_multiline_extends_list = false
+ij_kotlin_align_multiline_extends_list = true
 ij_kotlin_align_multiline_method_parentheses = false
 ij_kotlin_align_multiline_parameters = true
-ij_kotlin_align_multiline_parameters_in_calls = false
+ij_kotlin_align_multiline_parameters_in_calls = true
 ij_kotlin_allow_trailing_comma = false
 ij_kotlin_allow_trailing_comma_on_call_site = false
 ij_kotlin_assignment_wrap = normal
@@ -881,16 +882,17 @@ ij_kotlin_call_parameters_right_paren_on_new_line = true
 ij_kotlin_call_parameters_wrap = on_every_item
 ij_kotlin_catch_on_new_line = false
 ij_kotlin_class_annotation_wrap = split_into_lines
-ij_kotlin_continuation_indent_for_chained_calls = true
-ij_kotlin_continuation_indent_for_expression_bodies = true
-ij_kotlin_continuation_indent_in_argument_lists = true
-ij_kotlin_continuation_indent_in_elvis = true
-ij_kotlin_continuation_indent_in_if_conditions = true
-ij_kotlin_continuation_indent_in_parameter_lists = true
-ij_kotlin_continuation_indent_in_supertype_lists = true
+ij_kotlin_code_style_defaults = KOTLIN_OFFICIAL
+ij_kotlin_continuation_indent_for_chained_calls = false
+ij_kotlin_continuation_indent_for_expression_bodies = false
+ij_kotlin_continuation_indent_in_argument_lists = false
+ij_kotlin_continuation_indent_in_elvis = false
+ij_kotlin_continuation_indent_in_if_conditions = false
+ij_kotlin_continuation_indent_in_parameter_lists = false
+ij_kotlin_continuation_indent_in_supertype_lists = false
 ij_kotlin_else_on_new_line = false
-ij_kotlin_enum_constants_wrap = off
-ij_kotlin_extends_list_wrap = normal
+ij_kotlin_enum_constants_wrap = normal
+ij_kotlin_extends_list_wrap = on_every_item
 ij_kotlin_field_annotation_wrap = split_into_lines
 ij_kotlin_finally_on_new_line = false
 ij_kotlin_if_rparen_on_new_line = true
@@ -909,13 +911,13 @@ ij_kotlin_line_comment_add_space = false
 ij_kotlin_line_comment_add_space_on_reformat = false
 ij_kotlin_line_comment_at_first_column = true
 ij_kotlin_method_annotation_wrap = split_into_lines
-ij_kotlin_method_call_chain_wrap = normal
+ij_kotlin_method_call_chain_wrap = on_every_item
 ij_kotlin_method_parameters_new_line_after_left_paren = true
 ij_kotlin_method_parameters_right_paren_on_new_line = true
 ij_kotlin_method_parameters_wrap = on_every_item
 ij_kotlin_name_count_to_use_star_import = 2147483647
 ij_kotlin_name_count_to_use_star_import_for_members = 2147483647
-ij_kotlin_parameter_annotation_wrap = off
+ij_kotlin_parameter_annotation_wrap = split_into_lines
 ij_kotlin_space_after_comma = true
 ij_kotlin_space_after_extend_colon = true
 ij_kotlin_space_after_type_colon = true
@@ -930,7 +932,7 @@ ij_kotlin_space_before_when_parentheses = true
 ij_kotlin_space_before_while_parentheses = true
 ij_kotlin_spaces_around_additive_operators = true
 ij_kotlin_spaces_around_assignment_operators = true
-ij_kotlin_spaces_around_equality_operators = true
+ij_kotlin_spaces_around_equality_operators = false
 ij_kotlin_spaces_around_function_type_arrow = true
 ij_kotlin_spaces_around_logical_operators = true
 ij_kotlin_spaces_around_multiplicative_operators = true
@@ -939,7 +941,7 @@ ij_kotlin_spaces_around_relational_operators = true
 ij_kotlin_spaces_around_unary_operator = false
 ij_kotlin_spaces_around_when_arrow = true
 ij_kotlin_use_custom_formatting_for_modifiers = true
-ij_kotlin_variable_annotation_wrap = off
+ij_kotlin_variable_annotation_wrap = normal
 ij_kotlin_while_on_new_line = false
 ij_kotlin_wrap_elvis_expressions = 1
 ij_kotlin_wrap_expression_body_functions = 1
