+++ frontend/app/src/androidTest/kotlin/com/example/calendy/ScheduleDaoTest.kt
@@ -3,16 +3,16 @@ package com.example.calendy
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.example.calendy.data.schedule.ScheduleDao
+import com.example.calendy.data.plan.schedule.ScheduleDao
 import org.junit.Before
 import org.junit.runner.RunWith
 import androidx.room.Room
 import com.example.calendy.data.CalendyDatabase
 import org.junit.After
 import java.io.IOException
-import com.example.calendy.data.schedule.Schedule
-import com.example.calendy.data.schedule.ScheduleLocalDataSource
-import com.example.calendy.data.schedule.ScheduleRepository
+import com.example.calendy.data.plan.Schedule
+import com.example.calendy.data.plan.schedule.ScheduleLocalDataSource
+import com.example.calendy.data.plan.schedule.ScheduleRepository
 import junit.framework.TestCase.assertEquals
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
