+++ frontend/app/src/main/java/com/example/calendy/data/CalendyApi.kt
@@ -1,7 +1,5 @@
 package com.example.calendy.data
 
-import com.example.calendy.data.schedule.Schedule
-import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
