+++ frontend/app/src/main/java/com/example/calendy/data/plan/Plan.kt
@@ -1,0 +1,12 @@
+package com.example.calendy.data.plan
+
+sealed interface Plan {
+    val id: Int
+    val title: String
+    val memo: String
+    val categoryId: Int?
+    val repeatGroupId: Int?
+    val priority: Int
+    val showInMonthlyView: Boolean
+    val isOverridden: Boolean
+}
