+++ frontend/app/src/main/java/com/example/calendy/data/plan/schedule/IScheduleRepository.kt
@@ -1,6 +1,6 @@
-package com.example.calendy.data.schedule
+package com.example.calendy.data.plan.schedule
 
-import com.example.calendy.data.schedule.Schedule
+import com.example.calendy.data.plan.Schedule
 import kotlinx.coroutines.flow.Flow
 import java.util.Date
