+++ frontend/app/src/main/java/com/example/calendy/data/plan/schedule/ScheduleLocalDataSource.kt
@@ -1,5 +1,6 @@
-package com.example.calendy.data.schedule
+package com.example.calendy.data.plan.schedule
 
+import com.example.calendy.data.plan.Schedule
 import kotlinx.coroutines.flow.Flow
 import java.util.Date
