+++ frontend/app/src/main/java/com/example/calendy/data/plan/schedule/ScheduleRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package com.example.calendy.data.schedule
+package com.example.calendy.data.plan.schedule
 
 import com.example.calendy.data.CalendyApi
