+++ frontend/app/src/main/java/com/example/calendy/data/plan/todo/ITodoRepository.kt
@@ -1,6 +1,6 @@
-package com.example.calendy.data.todo
+package com.example.calendy.data.plan.todo
 
-import com.example.calendy.data.todo.Todo
+import com.example.calendy.data.plan.Todo
 import kotlinx.coroutines.flow.Flow
 import java.util.Date
