+++ frontend/app/src/main/java/com/example/calendy/view/editplanview/EditPlanPage.kt
@@ -68,7 +68,7 @@ import com.example.calendy.data.DummyCategoryRepository
 import com.example.calendy.data.DummyScheduleRepository
 import com.example.calendy.data.DummyTodoRepository
 import com.example.calendy.data.category.Category
-import com.example.calendy.data.schedule.ScheduleRepository
+import com.example.calendy.data.plan.schedule.ScheduleRepository
 import com.gowtham.ratingbar.RatingBar
 import com.gowtham.ratingbar.RatingBarConfig
 import com.holix.android.bottomsheetdialog.compose.BottomSheetDialog
